telegram_service/requirements.txt
python-telegram-bot[webhooks]==20.8
python-dotenv
requests
aiohttp
openai